Add post method

+++ pages/api/products/[id].jsx
@@ -5,7 +5,9 @@ export default async function handler(req, res) {
     const {
         method,
         query: {id},
+        cookies,
     } = req;
+    const token = cookies.token;
     dbConnect();
     if (method === 'GET') {
         try {
@@ -16,14 +18,22 @@ export default async function handler(req, res) {
         }
     }
     if (method === 'PUT') {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json('Not Authenticated');
+        }
         try {
-            const product = await Product.create(req.body);
+            const product = await Product.findByIdAndUpdate(id, req.body, {
+                new: true,
+            });
             res.status(201).json(product);
         } catch (error) {
             res.status(500).json(error);
         }
     }
     if (method === 'DELETE') {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json('Not Authenticated');
+        }
         try {
             await Product.findByIdAndDelete(id);
             res.status(201).json('The Product has been deleted!');
